feat: Integrate NGINX masters with updated Docker configurations and FastAPI routing
- Refactor master nodes to use new Dockerfile and environment variables for worker bootstrapping
- Implement master router for handling API requests
- Remove mock API and replace with master router logic
- Update onboarding documentation to reflect changes in architecture
- Adjust inference model loading to support bfloat16 on CPU

+++ workers/inference.py
@@ -38,7 +38,7 @@ def load_model(model_path: str, device: str = "cuda:0"):
 
     _model = AutoModelForCausalLM.from_pretrained(
         model_path,
-        torch_dtype=torch.float16 if "cuda" in _device else torch.float32,
+        torch_dtype=torch.float16 if "cuda" in _device else torch.bfloat16,
     ).to(_device)
     _model.eval()
